NONE: refactor prometheus integration
