+++ core/internal/http/http_metrics.go
@@ -1,12 +1,12 @@
-package metrics
+package http
 
 import (
 	"github.com/gin-gonic/gin"
 	ginprometheus "github.com/zsais/go-gin-prometheus"
 )
 
-// ApplyRoutes workspace route handlers
-func ApplyRoutes(r *gin.Engine) {
+// ApplyMetrics apply metrics to routes
+func ApplyMetrics(r *gin.Engine) {
 	p := ginprometheus.NewPrometheus("gin")
 	p.Use(r)
 }
