+++ core/prometheus.yml
@@ -1,6 +1,6 @@
 scrape_configs:
-    - job_name: myapp
-      scrape_interval: 10s
-      static_configs:
-      - targets:
-        - localhost:5051
+  - job_name: flagbase-core
+    scrape_interval: 10s
+    static_configs:
+    - targets:
+      - localhost:5051
